Moved MainMenuPatcher to correct namespace, fixed typo
