+++ src/SpaceWarp.UI/Patching/MainMenu/MainMenuPatcher.cs
@@ -3,11 +3,10 @@ using I2.Loc;
 using JetBrains.Annotations;
 using KSP.Api.CoreTypes;
 using KSP.Game.StartupFlow;
-using SpaceWarp.API.UI;
 using TMPro;
 using UnityObject = UnityEngine.Object;
 
-namespace SpaceWarp.Patching.Settings;
+namespace SpaceWarp.Patching.MainMenu;
 
 [HarmonyPatch(typeof(LandingHUD), nameof(LandingHUD.Start))]
 internal class MainMenuPatcher
@@ -21,7 +20,7 @@ internal class MainMenuPatcher
         var menuItemsGroupTransform = __instance.transform.FindChildEx("MenuItemsGroup");
         var singleplayerButtonTransform = menuItemsGroupTransform.FindChildEx("Singleplayer");
 
-        foreach (var menuButtonToBeAdded in MainMenu.MenuButtonsToBeAdded)
+        foreach (var menuButtonToBeAdded in API.UI.MainMenu.MenuButtonsToBeAdded)
         {
             var newButton =
                 UnityObject.Instantiate(singleplayerButtonTransform.gameObject, menuItemsGroupTransform, false);
@@ -44,10 +43,14 @@ internal class MainMenuPatcher
             LocalizationManager.OnLocalizeEvent += () => tmp.SetText(menuButtonToBeAdded.name);
         }
 
-        foreach (var localizedMenuButtonToBeAddded in MainMenu.LocalizedMenuButtonsToBeAdded)
-        {var newButton =
-                UnityObject.Instantiate(singleplayerButtonTransform.gameObject, menuItemsGroupTransform, false);
-            newButton.name = localizedMenuButtonToBeAddded.term;
+        foreach (var localizedMenuButtonToBeAdded in API.UI.MainMenu.LocalizedMenuButtonsToBeAdded)
+        {
+            var newButton = UnityObject.Instantiate(
+                singleplayerButtonTransform.gameObject,
+                menuItemsGroupTransform,
+                false
+            );
+            newButton.name = localizedMenuButtonToBeAdded.term;
 
             // Move the button to be above the Exit button.
             newButton.transform.SetSiblingIndex(newButton.transform.GetSiblingIndex() - 1);
@@ -55,11 +58,12 @@ internal class MainMenuPatcher
             // Rebind the button's action to call the action
             var uiAction = newButton.GetComponent<UIAction_Void_Button>();
             DelegateAction action = new();
-            action.BindDelegate(() => localizedMenuButtonToBeAddded.onClicked.Invoke());
+            action.BindDelegate(() => localizedMenuButtonToBeAdded.onClicked.Invoke());
             uiAction.BindAction(action);
             var localize = newButton.GetComponentInChildren<Localize>();
-            localize.SetTerm(localizedMenuButtonToBeAddded.term);
+            localize.SetTerm(localizedMenuButtonToBeAdded.term);
         }
+
         MainMenuLoaded?.Invoke();
     }
 }
